refactor: http와 controller관련 클래스의 패키지 구조 변경
- HttpUtils에서 stream 대신 for문으로 변경
- HttpUtils에서 메서드명 변경
