+++ src/main/java/controller/FrontController.java
@@ -1,8 +1,11 @@
 package controller;
 
 import controller.annotation.CustomRequestMapping;
-import model.CustomHttpRequest;
-import model.CustomHttpResponse;
+import controller.controllers.BaseController;
+import controller.controllers.UserController;
+import controller.controllers.ViewController;
+import http.CustomHttpRequest;
+import http.CustomHttpResponse;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
