+++ src/main/java/controller/controllers/BaseController.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.controllers;
 
 public abstract class BaseController {
 }
