+++ src/main/java/controller/controllers/ViewController.java
@@ -1,9 +1,9 @@
-package controller;
+package controller.controllers;
 
 import controller.annotation.CustomRequestMapping;
-import model.CustomHttpMethod;
-import model.CustomHttpRequest;
-import model.CustomHttpResponse;
+import http.CustomHttpMethod;
+import http.CustomHttpRequest;
+import http.CustomHttpResponse;
 import utils.FileIoUtils;
 
 import java.util.HashMap;
