+++ src/main/java/http/CustomHttpMethod.java
@@ -1,4 +1,4 @@
-package model;
+package http;
 
 import java.util.Arrays;
