+++ src/main/java/http/CustomHttpRequest.java
@@ -1,4 +1,4 @@
-package model;
+package http;
 
 import java.util.Map;
