+++ src/main/java/http/CustomHttpResponse.java
@@ -1,4 +1,4 @@
-package model;
+package http;
 
 import java.util.Map;
